Restructure configuration a single file separate from the code
Put all user configuration in a single file to keep it separate from the code.
This also makeas it clearer that the configuration for groups and users are different.

This makes it easier to for end users when they need to update, just to update the
scripts and can levave the configuration unchanged.

Update the docmentation to reflect the changes.

Also add .gitignore with files that should not be checked in to git repository,
Inställningar.gs and files created when using the clasp tool locally.